move models from routes to /models
